Version 4.0
All projects fixed to build in both debug and release configurations.

Version and date information removed from all files. Version constants
added to SystemConfiguration.h in SKA library. Version identification
now assumed to happed at Git repository level.  Version constants were

App2001 project split into MotionDescriptor and PTAnalysis projects.

Visual Studio build configuration moved to VS 14.

Linux makefiles are out of date and likely broken.

Minor changes to SKA library:
Animation/Skeleton.cpp - some changes in Dec 2016
Camera/Camera.cpp - Modified limits on roll in move() method.
Core/SystemConfiguration.h - added #define constants with SKA version
numbers
Core/SystemLog.cpp - write SKA version to system log when it is opened
DataManagement/BVH_Reader.cpp - added some explicit casts to avoid
compiler warnings
Math/Matrix4x4.h - correction to Matrix4x4::multVector()
Math/Vector3D.h - added ndot() function for normalized dot product
Models/CodedModels.{h,cpp} - Added XYPlaneModel, YZPlaneModel and
ZXPlaneModel
Models/Models.h - Fixed alpha in ModelFace to be part of Color attribute
Models/ModelFactory.cpp - Added Vector into the model builder cases

This commit contains the version of the PTAnalysis project that was
delivered to Tony Laurel in February 2017.

The new MotionDescriptors project is currently just a place holder
(filled with same code as PTAnalysis). This project will be populated
once Trevor Martin has code to commit.

+++ SKA/SKA/src/Animation/MotionSequence.cpp
@@ -12,8 +12,7 @@
 // being credited for any significant use, particularly if used for
 // commercial projects or academic research publications.
 //-----------------------------------------------------------------------------
-// Version 3.1 - September 1, 2014 - Michael Doherty
-//-----------------------------------------------------------------------------
+
 #include <Core/SystemConfiguration.h>
 #include <complex>
 #include <Core/SystemLog.h>
@@ -35,13 +34,13 @@ public:
 		short i = getChannelIndexFromID(c);
 		if (i>=0) return i;
 		// add new channel, with a new index
-		short new_index = channels.size();
+		short new_index = short(channels.size());
 		channels.insert(pair<CHANNEL_ID,short>(c,new_index)); 
 		return new_index;
 	}
 	short numChannels() 
 	{ 
-		return channels.size(); 
+		return short(channels.size()); 
 	}
 	void clear()
 	{
@@ -75,7 +74,7 @@ public:
 		}
 		return i;
 	}
-	short size() { return channels.size(); }
+	short size() { return short(channels.size()); }
 };
 
 ostream& operator<<(ostream& out, ChannelMap& cm)
